cli: Improve CLI.
