+++ src/main.rs
@@ -24,18 +24,12 @@ mod errors;
 mod lexer;
 mod parser;
 mod vm;
+mod cli;
 
 // mod builtins;
 
-use crate::{
-    compiler::Compiler,
-    lexer::Lexer,
-    parser::Parser,
-    vm::{Value, VM},
-};
-use clap::{App, Arg};
+use crate::cli::cli;
 pub use errors::{OrionError, Result};
-use rustyline::{error::ReadlineError, Editor};
 use std::{fs, path::Path, process::exit, time::Instant};
 
 #[macro_export]
@@ -82,215 +76,9 @@ fn print_err(e: OrionError) {
     );
 }
 
-fn repl(no_prelude: bool, debug: bool, quiet: bool) -> Result<()> {
-    println!(
-        ";; Orion REPL v{}.\n
-;; Copyright (C) 2021  Wafelack <wafelack@protonmail.com>
-;; This program comes with ABSOLUTELY NO WARRANTY.
-;; This is free software, and you are welcome to redistribute it
-;; under certain conditions.",
-        env!("CARGO_PKG_VERSION")
-    );
-    // let mut interpreter = Interpreter::new(vec![], no_prelude, quiet)?;
-
-    let mut rl = Editor::<()>::new();
-    let mut i = 0;
-
-    loop {
-        i += 1;
-        let line = rl.readline("orion> ");
-
-        match line {
-            Ok(line) => {
-                rl.add_history_entry(line.as_str());
-                if line == "(quit)" {
-                    return Ok(());
-                }
-
-                let tokens = match Lexer::new(line.trim(), "REPL").line(i).proc_tokens() {
-                    Ok(t) => t,
-                    Err(e) => {
-                        print_err(e);
-                        continue;
-                    }
-                };
-
-                if debug {
-                    println!("Tokens\n======");
-                    tokens.iter().for_each(|t| {
-                        println!("{}", t.display());
-                    });
-                    println!();
-                }
-
-                let ast = match Parser::new(tokens, "REPL").parse() {
-                    Ok(ast) => ast,
-                    Err(e) => {
-                        print_err(e);
-                        continue;
-                    }
-                };
-
-                if debug {
-                    println!("Syntax Tree\n===========");
-                    ast.iter().for_each(|e| println!("{}", e.get_type()));
-                }
-
-                // interpreter.update_ast(ast);
-
-                if debug {
-                    println!("\nStdout\n======");
-                }
-
-                let bytecode = Compiler::new(ast, "REPL").compile()?;
-                println!("[INSTRUCTIONS]");
-                bytecode
-                    .instructions
-                    .iter()
-                    .for_each(|i| println!("{:?}", i));
-                println!("\n[SYMBOLS]");
-                bytecode
-                    .symbols
-                    .iter()
-                    .enumerate()
-                    .for_each(|(idx, var)| println!("0x{:04x}: {}", idx, var));
-                println!("\n[CONSTANTS]");
-                bytecode
-                    .constants
-                    .iter()
-                    .enumerate()
-                    .for_each(|(idx, constant)| println!("0x{:04x}: {:?}", idx, constant));
-                println!("\n[CHUNKS]");
-                bytecode.chunks.iter().enumerate().for_each(|(idx, chunk)| {
-                    println!("{}: {{", idx);
-                    println!("  reference: [");
-                    chunk.reference.iter().for_each(|sym| {
-                        println!("    0x{:04x}", sym);
-                    });
-                    println!("  ]\n  instructions: [");
-                    chunk.instructions.iter().for_each(|instr| {
-                        println!("    {:?}", instr);
-                    });
-                    println!("  ]\n}}");
-                });
-
-                let mut vm = VM::<256>::new(bytecode);
-                println!("-------------------------------");
-                println!("[STDOUT]");
-                let memory = vm.eval()?;
-                println!("\n[STACK]");
-                vm.stack.into_iter().enumerate().for_each(|(idx, v)| {
-                    println!("0x{:02x}: {:?}", idx, v);
-                });
-                println!("\n[MEMORY]");
-                memory.into_iter().enumerate().for_each(|(idx, v)| {
-                    println!("0x{:02x}: {:?}", idx, v);
-                });
-
-                /*
-                   let start = Instant::now();
-                /* match interpreter.interpret(true) {
-                Ok(_) => {},
-                Err(e) => {
-                print_err(e);
-                continue;
-                }
-                } */
-                let elapsed = start.elapsed();
-                if debug {
-                    println!("\nDone in {}ms.", elapsed.as_millis());
-                } */
-            }
-            Err(ReadlineError::Interrupted) => {
-                println!(";; User break");
-            }
-            Err(ReadlineError::Eof) => return Ok(()),
-            Err(_) => {
-                eprintln!("An error occured, please retry.");
-                continue;
-            }
-        }
-    }
-}
-
-fn try_main() -> Result<()> {
-    let matches = App::new("orion")
-        .author(env!("CARGO_PKG_AUTHORS"))
-        .version(env!("CARGO_PKG_VERSION"))
-        .about("Orion is a purely functional lisp dialect.")
-        .arg(
-            Arg::with_name("file")
-                .takes_value(true)
-                .index(1)
-                .help("The source file to pass to the interpreter"),
-        )
-        .arg(
-            Arg::with_name("no-load-prelude")
-                .short("np")
-                .long("no-load-prelude")
-                .help("Do not load the prelude file"),
-        )
-        .arg(
-            Arg::with_name("debug")
-                .short("d")
-                .long("debug")
-                .help("Display debug information."),
-        )
-        .arg(
-            Arg::with_name("quiet")
-                .short("q")
-                .long("quiet")
-                .help("Do not display messages."),
-        )
-        .get_matches();
-
-    if let Some(path) = matches.value_of("file") {
-        if Path::new(path).exists() {
-            let content = match fs::read_to_string(path) {
-                Ok(c) => c,
-                Err(e) => return error!(=> "fatal: Failed to read file: {}.", e),
-            };
-
-            let tokens = Lexer::new(content, path).proc_tokens()?;
-            if matches.is_present("debug") {
-                println!("Tokens\n======");
-                tokens.iter().for_each(|t| {
-                    println!("{}", t.display());
-                });
-                println!();
-            }
-
-            let ast = Parser::new(tokens, path).parse()?;
-
-            if matches.is_present("debug") {
-                println!("Syntax Tree\n===========");
-                ast.iter().for_each(|e| println!("{}", e.get_type()));
-            }
-            if matches.is_present("debug") {
-                println!("\nStdout\n======");
-            }
-            let start = Instant::now();
-            // Interpreter::new(ast, matches.is_present("no-load-prelude"), matches.is_present("quiet"))?.interpret(false)?;
-            let elapsed = start.elapsed();
-            if matches.is_present("debug") {
-                println!("\nDone in {}ms.", elapsed.as_millis());
-            }
-            Ok(())
-        } else {
-            error!(=> "fatal: File not found: {}.", path)
-        }
-    } else {
-        repl(
-            matches.is_present("no-load-prelude"),
-            matches.is_present("debug"),
-            matches.is_present("quiet"),
-        )?;
-        Ok(())
-    }
-}
 
 fn main() {
-    match try_main() {
+    match cli() {
         Ok(()) => {}
         Err(e) => {
             print_err(e);
